ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
opencv-python-headless==4.9.0.80
Pillow==10.3.0
exifread==3.0.0
transformers==4.41.2
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
requests==2.31.0
